Add battle persistence; service, mappers, controller and repository for battle
