+++ src/main/java/ru/redsoft/seabattles/WarshipsApplication.java
@@ -1,0 +1,13 @@
+package ru.redsoft.seabattles;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WarshipsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WarshipsApplication.class, args);
+    }
+
+}
