+++ src/main/java/ru/redsoft/seabattles/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package ru.redsoft.seabattles.exception;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
