+++ src/main/java/ru/redsoft/seabattles/exchange/response/CountryResponse.java
@@ -1,0 +1,19 @@
+package ru.redsoft.seabattles.exchange.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import ru.redsoft.seabattles.persistence.Side;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class CountryResponse {
+
+    private String name;
+
+    private Side side;
+
+}
