+++ src/main/java/ru/redsoft/seabattles/persistence/Battle.java
@@ -1,0 +1,20 @@
+package ru.redsoft.seabattles.persistence;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.Instant;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Battle {
+
+    private String battleName;
+
+    private Instant battleDate;
+
+}
