+++ src/main/java/ru/redsoft/seabattles/persistence/Country.java
@@ -1,0 +1,17 @@
+package ru.redsoft.seabattles.persistence;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Country {
+
+    private String name;
+
+    private Side side;
+}
