+++ src/main/java/ru/redsoft/seabattles/persistence/Side.java
@@ -1,0 +1,7 @@
+package ru.redsoft.seabattles.persistence;
+
+public enum Side {
+
+    ALLIES,
+    AXIS
+}
