+++ src/main/java/ru/redsoft/seabattles/persistence/WarshipClass.java
@@ -1,0 +1,12 @@
+package ru.redsoft.seabattles.persistence;
+
+public enum WarshipClass {
+
+    BATTLESHIP,
+    AIRCRAFT_CARRIER,
+    CRUISER,
+    DESTROYER,
+    SUBMARINE,
+    MERCHANT_SHIP
+
+}
