+++ src/main/java/ru/redsoft/seabattles/repository/CountryRepository.java
@@ -1,0 +1,10 @@
+package ru.redsoft.seabattles.repository;
+
+import ru.redsoft.seabattles.persistence.Country;
+
+import java.util.List;
+
+public interface CountryRepository {
+
+    List<Country> findAll();
+}
