+++ src/main/java/ru/redsoft/seabattles/repository/WarshipRepository.java
@@ -1,0 +1,13 @@
+package ru.redsoft.seabattles.repository;
+
+import ru.redsoft.seabattles.persistence.Warship;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface WarshipRepository {
+
+    Optional<Warship> findById(Long id);
+
+    List<Warship> findAll();
+}
