--- src/main/java/ru/redsoft/warships/WarshipsApplication.java
@@ -1,13 +1,0 @@
-package ru.redsoft.warships;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WarshipsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(WarshipsApplication.class, args);
-    }
-
-}
