--- src/main/java/ru/redsoft/warships/controller/CountryRestController.java
@@ -1,24 +1,0 @@
-package ru.redsoft.warships.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import ru.redsoft.warships.exchange.response.CountryResponse;
-import ru.redsoft.warships.service.CountryService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/v1/countries")
-@RequiredArgsConstructor
-public class CountryRestController {
-
-    private final CountryService countryService;
-
-    @GetMapping
-    public List<CountryResponse> findAll() {
-        return countryService.findAll();
-    }
-
-}
