--- src/main/java/ru/redsoft/warships/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package ru.redsoft.warships.exception;
-
-public class NotFoundException extends RuntimeException {
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
