--- src/main/java/ru/redsoft/warships/exchange/response/CountryResponse.java
@@ -1,19 +1,0 @@
-package ru.redsoft.warships.exchange.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import ru.redsoft.warships.persistence.Side;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class CountryResponse {
-
-    private String name;
-
-    private Side side;
-
-}
