--- src/main/java/ru/redsoft/warships/persistence/Country.java
@@ -1,17 +1,0 @@
-package ru.redsoft.warships.persistence;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Country {
-
-    private String name;
-
-    private Side side;
-}
