--- src/main/java/ru/redsoft/warships/persistence/Side.java
@@ -1,7 +1,0 @@
-package ru.redsoft.warships.persistence;
-
-public enum Side {
-
-    ALLIES,
-    AXIS
-}
