--- src/main/java/ru/redsoft/warships/persistence/WarshipClass.java
@@ -1,12 +1,0 @@
-package ru.redsoft.warships.persistence;
-
-public enum WarshipClass {
-
-    BATTLESHIP,
-    AIRCRAFT_CARRIER,
-    CRUISER,
-    DESTROYER,
-    SUBMARINE,
-    MERCHANT_SHIP
-
-}
