--- src/main/java/ru/redsoft/warships/repository/CountryRepository.java
@@ -1,10 +1,0 @@
-package ru.redsoft.warships.repository;
-
-import ru.redsoft.warships.persistence.Country;
-
-import java.util.List;
-
-public interface CountryRepository {
-
-    List<Country> findAll();
-}
