+++ src/test/java/ru/redsoft/seabattles/WarshipApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.redsoft.seabattles;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WarshipApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
