--- src/test/java/ru/redsoft/warships/WarshipApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.redsoft.warships;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WarshipApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
